Running test with the new approximation of the new effects.
This test is intended to go with the master branch where I am using the new approximation to calculate the electrostatic bremsstrahlung.

The idea is doing the same analysis that was done in the ApJL paper, but using the new expressions for both collisional damping and electrostatic bremsstrahlung. So I am not including both effects in the initial spectrum for L and S waves.

In "tevol" I am doing an analysis with EB and CD included in the initial L and S waves spectra.

I love git and git integration for emacs! Makes doing science easier! =D